add media processing to fetcher, add emoji to tag
